requirements.txt
Flask
pymongo[srv]
Flask-PyMongo
python-dotenv
gunicorn
flask-cors